+++ internal/api/v1/logout/handler.go
@@ -1,0 +1,16 @@
+package logout
+
+import (
+	"net/http"
+)
+
+type Handler struct {
+	storage storageRepository
+}
+
+func New(storage storageRepository) *Handler {
+	return &Handler{storage: storage}
+}
+
+func (h *Handler) ServeHTTP(resp http.ResponseWriter, req *http.Request) {
+}