+++ internal/api/v1/logout/request.go
@@ -1,0 +1,5 @@
+package logout
+
+type request struct {
+	Token string `json:"token"`
+}