+++ internal/api/v1/logout/storage.go
@@ -1,0 +1,9 @@
+package logout
+
+import (
+	"context"
+)
+
+type storageRepository interface {
+	Logout(ctx context.Context, token string) error
+}